remove network scope and move commands to browser
