+++ packages/wdio-utils/src/monad.js
@@ -35,6 +35,8 @@ export default function WebDriver (options, modifier, propertiesObject = {}) {
         propertiesObject.commandList = { value: Object.keys(propertiesObject) }
         propertiesObject.options = { value: options }
         propertiesObject.requestedCapabilities = { value: options.requestedCapabilities }
+        propertiesObject.puppeteer = { value: null, writable: true }
+        propertiesObject._NOT_FIBER = { value: false, writable: true }
 
         /**
          * allow to wrap commands if necessary
