+++ packages/webdriverio/src/commands/browser.js
@@ -10,6 +10,7 @@ export const executeAsync = require('./browser/executeAsync').default
 export const getCookies = require('./browser/getCookies').default
 export const getWindowSize = require('./browser/getWindowSize').default
 export const keys = require('./browser/keys').default
+export const mock = require('./browser/mock').default
 export const newWindow = require('./browser/newWindow').default
 export const pause = require('./browser/pause').default
 export const react$$ = require('./browser/react$$').default
@@ -21,6 +22,7 @@ export const setCookies = require('./browser/setCookies').default
 export const setTimeout = require('./browser/setTimeout').default
 export const setWindowSize = require('./browser/setWindowSize').default
 export const switchWindow = require('./browser/switchWindow').default
+export const throttle = require('./browser/throttle').default
 export const touchAction = require('./browser/touchAction').default
 export const uploadFile = require('./browser/uploadFile').default
 export const url = require('./browser/url').default
