+++ packages/webdriverio/src/commands/browser/mock.js
@@ -18,18 +18,18 @@
     :mock.js
     it('should mock network resources', () => {
         // via static string
-        const userListMock = browser.network.mock('**' + '/users/list')
+        const userListMock = browser.mock('**' + '/users/list')
         // you can also specifying the mock even more by filtering resources
         // by headers or status code, e.g. mock only responses with specific
         // header set
-        const strictMock = browser.network.mock('**', {
+        const strictMock = browser.mock('**', {
             // mock all json responses
             headers: { 'Content-Type': 'application/json' }
         })
     })
 
     it('should modify API responses', () => {
-        const todoMock = browser.network.mock('**' + '/todos', {
+        const todoMock = browser.mock('**' + '/todos', {
             method: 'get'
         })
 
@@ -56,15 +56,15 @@
     })
 
     it('should modify text assets', () => {
-        const scriptMock = browser.network.mock('**' + '/script.min.js')
+        const scriptMock = browser.mock('**' + '/script.min.js')
         scriptMock.respond('./tests/fixtures/script.js')
     })
 
     it('should redirect web resources', () => {
-        const headerMock = browser.network.mock('**' + '/header.png')
+        const headerMock = browser.mock('**' + '/header.png')
         headerMock.respond('https://media.giphy.com/media/F9hQLAVhWnL56/giphy.gif')
 
-        const pageMock = browser.network.mock('https://google.com/')
+        const pageMock = browser.mock('https://google.com/')
         pageMock.respond('https://webdriver.io')
         browser.url('https://google.com')
         console.log(browser.getTitle()) // returns "WebdriverIO · Next-gen browser and mobile automation test framework for Node.js"
@@ -80,12 +80,16 @@
  */
 import DevtoolsNetworkInterception from '../../utils/interception/devtools'
 import WebDriverNetworkInterception from '../../utils/interception/webdriver'
-import { getBrowserObject } from '../../utils'
+import { getBrowserObject, getPuppeteer } from '../../utils'
 
 const SESSION_MOCKS = new Set()
 
 export default async function mock (url, filterOptions) {
     const NetworkInterception = this.isSauce ? WebDriverNetworkInterception : DevtoolsNetworkInterception
+
+    if (!this.isSauce) {
+        await getPuppeteer.call(this)
+    }
 
     /**
      * enable network Mocking if not already
