+++ packages/webdriverio/src/commands/mock/respond.js
@@ -6,7 +6,7 @@
  * <example>
     :respond.js
     it('should demonstrate the addValue command', () => {
-        const mock = browser.network.mock('https://todo-backend-express-knex.herokuapp.com/', {
+        const mock = browser.mock('https://todo-backend-express-knex.herokuapp.com/', {
             method: 'get'
         })
