--- packages/webdriverio/src/commands/network.js
@@ -1,2 +1,0 @@
-export const mock = require('./network/mock').default
-export const throttle = require('./network/throttle').default
