--- packages/webdriverio/tests/commands/network.test.js
@@ -1,16 +1,0 @@
-import fs from 'fs'
-import path from 'path'
-import { getPrototype } from '../../src/utils'
-
-const scope = 'network'
-const dir = path.resolve(__dirname, '../..', 'src', 'commands', scope)
-const files = fs
-    .readdirSync(dir)
-    .map(f => path.basename(f, path.extname(f)))
-
-test(scope + ' commands list and strategies', () => {
-    const prototype = Object.keys(getPrototype(scope))
-    const expected = [...files, 'strategies']
-
-    expect(prototype).toEqual(expected)
-})
