+++ packages/webdriverio/tests/utils.test.js
@@ -1,6 +1,8 @@
 import fs from 'fs'
 import http from 'http'
 import path from 'path'
+import puppeteer from 'puppeteer-core'
+
 import { ELEMENT_KEY } from '../src/constants'
 import {
     getElementFromResponse,
@@ -16,6 +18,7 @@ import {
     assertDirectoryExists,
     validateUrl,
     getAutomationProtocol,
+    getPuppeteer,
     updateCapabilities
 } from '../src/utils'
 
@@ -33,7 +36,18 @@ jest.mock('http', () => {
     }
 })
 
+const browser = {
+    sessionId: '1234',
+    capabilities: { browserName: 'foobar' },
+    __propertiesObject__: {},
+    requestedCapabilities: { requested: 'capabilities' }
+}
+
 jest.mock('fs')
+
+beforeEach(() => {
+    puppeteer.connect.mockClear()
+})
 
 describe('utils', () => {
     describe('getElementFromResponse', () => {
@@ -517,6 +531,112 @@ describe('utils', () => {
             http.setResponse({ statusCode: 404 })
             expect(await getAutomationProtocol({ capabilities: { browserName: 'foobar' } }))
                 .toBe('webdriver')
+        })
+    })
+
+    describe('attach Puppeteer', () => {
+        it('should fail if capabilities are not supported', async () => {
+            const err = await getPuppeteer.call(browser).catch(err => err)
+            expect(err.message).toContain('Network primitives aren\'t available for this session.')
+        })
+
+        it('should pass for Chrome', async () => {
+            const pptr = await getPuppeteer.call({
+                ...browser,
+                capabilities: {
+                    browserName: 'chrome',
+                    'goog:chromeOptions': {
+                        debuggerAddress: 'localhost:1234'
+                    }
+                }
+            })
+            expect(typeof pptr).toBe('object')
+            expect(puppeteer.connect.mock.calls).toMatchSnapshot()
+        })
+
+        it('should pass for Firefox', async () => {
+            const pprt = await getPuppeteer.call({
+                ...browser,
+                capabilities: {
+                    browserName: 'firefox',
+                    browserVersion: '79.0b'
+                },
+                requestedCapabilities: {
+                    'moz:firefoxOptions': {
+                        args: ['foo', 'bar', '-remote-debugging-port', 4321, 'barfoo']
+                    }
+                }
+            })
+            expect(typeof pprt).toBe('object')
+            expect(puppeteer.connect.mock.calls).toMatchSnapshot()
+        })
+
+        it('should pass for Firefox (DevTools)', async () => {
+            const pptr = await getPuppeteer.call({
+                ...browser,
+                capabilities: {
+                    browserName: 'firefox',
+                    browserVersion: '79.0b',
+                    'moz:firefoxOptions': {
+                        debuggerAddress: 'localhost:1234'
+                    }
+                },
+                requestedCapabilities: {
+                    'moz:firefoxOptions': {
+                        args: {}
+                    }
+                }
+            })
+            expect(typeof pptr).toBe('object')
+            expect(puppeteer.connect.mock.calls).toMatchSnapshot()
+        })
+
+        it('should pass for Edge', async () => {
+            const pptr = await getPuppeteer.call({
+                ...browser,
+                capabilities: {
+                    browserName: 'edge',
+                    'ms:edgeOptions': {
+                        debuggerAddress: 'localhost:1234'
+                    }
+                },
+                options: {
+                    requestedCapabilities: {}
+                }
+            })
+            expect(typeof pptr).toBe('object')
+            expect(puppeteer.connect.mock.calls).toMatchSnapshot()
+        })
+
+        it('should fail for old Firefox version', async () => {
+            const err = await getPuppeteer.call({
+                ...browser,
+                capabilities: {
+                    browserName: 'firefox',
+                    browserVersion: '78.0b'
+                },
+                requestedCapabilities: {
+                    'moz:firefoxOptions': {
+                        args: ['foo', 'bar', '-remote-debugging-port', 4321, 'barfoo']
+                    }
+                }
+            }).catch(err => err)
+            expect(err.message).toContain('Network primitives aren\'t available for this session.')
+        })
+
+        it('should not re-attach if connection was already established', async () => {
+            const pptr = await getPuppeteer.call({
+                ...browser,
+                capabilities: {
+                    browserName: 'chrome',
+                    'goog:chromeOptions': {
+                        debuggerAddress: 'localhost:1234'
+                    }
+                },
+                puppeteer: 'foobar'
+            })
+            expect(pptr).toBe('foobar')
+            expect(puppeteer.connect).toHaveBeenCalledTimes(0)
         })
     })
